% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_bubble.R
\name{ccc_bubble}
\alias{ccc_bubble}
\title{bubble plot showing the detailed information of cell-cell communications}
\usage{
ccc_bubble(
  pfile,
  mfile,
  neg_log10_th = -log10(0.05),
  means_exp_log2_th = 1,
  notused.cell = NULL,
  used.cell = NULL,
  neg_log10_th2 = 3,
  means_exp_log2_th2 = c(-4, 6),
  cell.pair = NULL,
  gene.pair = NULL,
  color_palette = c("#313695", "#4575B4", "#ABD9E9", "#FFFFB3", "#FDAE61", "#F46D43",
    "#D73027", "#A50026"),
  text_size = 12
)
}
\arguments{
\item{pfile}{pvalues.txt文件路径}

\item{mfile}{means.txt文件路径}

\item{neg_log10_th}{前面画第一张图的时候用的什么值当作显著性阈值，这里保持一致}

\item{means_exp_log2_th}{表达均值也限制一下}

\item{notused.cell}{确定不包含的细胞，默认是空值}

\item{used.cell}{必须含有的细胞，默认是空值}

\item{neg_log10_th2}{限定显著性的最大值}

\item{means_exp_log2_th2}{限定表达值的范围}

\item{cell.pair}{这里是自定义的顺序，若是可选细胞对的子集，则只展示子集，若有交集则只展示交集；空值情况下，会根据可选细胞对自动排序}

\item{gene.pair}{作用同上}

\item{color_palette}{颜色编码向量，颜色由浅到深}

\item{text_size}{坐标轴文本的大小}
}
\value{
a ggplot2 object
}
\description{
bubble plot showing the detailed information of cell-cell communications
}
