% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_compare.R
\name{ccc_compare}
\alias{ccc_compare}
\title{compare the cellphonedb results of two groups}
\usage{
ccc_compare(
  group1.name = NULL,
  group2.name = NULL,
  group1.pfile = NULL,
  group1.mfile = NULL,
  group2.pfile = NULL,
  group2.mfile = NULL,
  p.threshold = 0.05,
  thre = 1,
  cell.pair = NULL,
  gene.pair = NULL,
  plot.width = NULL,
  plot.height = NULL,
  filename = ""
)
}
\arguments{
\item{group1.name}{第一组叫什么名字}

\item{group2.name}{第二组叫什么名字}

\item{group1.pfile}{第一组两个文件的路径}

\item{group2.pfile}{第二组两个文件的路径}

\item{thre}{类似log2FC}

\item{cell.pair}{想展示的细胞对}

\item{gene.pair}{想展示的gene对}

\item{filename}{}
}
\value{
a pdf file and a xlsx file
}
\description{
compare the cellphonedb results of two groups
}
