% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_compare2.R
\name{ccc_compare2}
\alias{ccc_compare2}
\title{compare the cellphonedb results of two groups after considering the L-R direction}
\usage{
ccc_compare2(
  group1.name = NULL,
  group2.name = NULL,
  group1.pfile = NULL,
  group1.mfile = NULL,
  group2.pfile = NULL,
  group2.mfile = NULL,
  p.threshold = 0.05,
  thre = 1,
  cell.pair = NULL,
  gene.pair = NULL,
  plot.width = NULL,
  plot.height = NULL,
  filename = ""
)
}
\arguments{
\item{group1.name}{第一组叫什么名字}

\item{group2.name}{第二组叫什么名字}

\item{group1.pfile}{第一组两个文件的路径}

\item{group2.pfile}{第二组两个文件的路径}

\item{thre}{类似log2FC}

\item{cell.pair}{想展示的细胞对}

\item{gene.pair}{想展示的gene对（注意：改变gene.pair会改变PDF的结果，但不会改变Excel；表格始终是gene.pair为null的结果，因为这个结果后续画别的图会用到，所以没让它变）}

\item{filename}{}
}
\value{
a pdf file and a xlsx file
}
\description{
compare the cellphonedb results of two groups after considering the L-R direction
}
