% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_line.R
\name{ccc_line}
\alias{ccc_line}
\title{compare the cellphonedb results of two groups using connecting plot}
\usage{
ccc_line(
  table.path = NULL,
  group1.name = NULL,
  group2.name = NULL,
  ligand.cell = NULL,
  receptor.cell = NULL,
  ligand.color = "#4dbbd6",
  receptor.color = "#90d1c1",
  pt.size = 6,
  line.thre1 = 0.5,
  line.thre2 = 6,
  file.name = "",
  plot.width = 25,
  plot.height = 20
)
}
\arguments{
\item{table.path}{the xlsx file from last step, ccc_compare2 function}

\item{ligand.cell}{产生ligand的细胞类型}

\item{receptor.cell}{产生receptor的细胞类型}

\item{line.thre1}{这两个参数可以控制线的（相对）粗细}

\item{plot.height}{}
}
\value{
a pdf file
}
\description{
compare the cellphonedb results of two groups using connecting plot
}
