% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_number_heatmap1.R
\name{ccc_number_heatmap1}
\alias{ccc_number_heatmap1}
\title{heatmap showing the number of interactions}
\usage{
ccc_number_heatmap1(
  pfile,
  pvalue.threshold = 0.05,
  order.of.celltype = NULL,
  ccc.number.max = NULL,
  size.of.text = c(18, 14, 14, 12),
  color.palette = c("#4393C3", "#ffdbba", "#B2182B")
)
}
\arguments{
\item{pfile}{the file path of "pvalues.txt"}

\item{ccc.number.max}{控制图中数值范围}

\item{size.of.text}{长度为4的数值向量，依次表示坐标轴title、text、图例title、图例text的文本大小}

\item{color.palette}{配色方案，长度为3的颜色编码向量}
}
\value{
a ggplot2 object
}
\description{
heatmap showing the number of interactions
}
