% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccc_number_line.R
\name{ccc_number_line}
\alias{ccc_number_line}
\title{connecting plot showing the total number of interactions}
\usage{
ccc_number_line(
  pfile,
  pvalue.threshold = 0.05,
  order.of.celltype = NULL,
  ccc.number.max = NULL,
  color.palette = NULL,
  color.line = NULL,
  width.parameter = 0.5,
  vertex.label.cex = 1,
  vertex.size = 30
)
}
\arguments{
\item{pfile}{the file path of "pvalues.txt"}

\item{ccc.number.max}{控制图中数值范围}

\item{color.palette}{节点的颜色，命名的颜色编码向量}

\item{color.line}{颜色编码向量，颜色深度需要递进}

\item{width.parameter}{调整连线的宽度}

\item{vertex.label.cex}{节点标注字体大小}

\item{vertex.size}{节点大小}
}
\description{
connecting plot showing the total number of interactions
}
