\name{CPplot-package}
\alias{CPplot-package}
\docType{package}
\title{Causal CP-Plot Diagnostics}
\description{
Implements CP-plots of estimated conditional average treatment effects against
estimated propensity scores. Motivated by covariance representations for
weighted average treatment effects, these plots summarize how treatment-effect
heterogeneity varies with treatment propensity and help diagnose bracketing
relationships among weighted causal estimands. The package also implements
local CP-plots for instrumental-variable analyses of weighted local average
treatment effects and accompanies the manuscript \emph{Introducing the CP-plot
Based on Covariance Representations of Weighted Average Treatment Effects}.
}
