% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpplot-package.R
\docType{package}
\name{cpplot-package}
\alias{cpplot}
\alias{cpplot-package}
\title{cpplot: Visualize the results of cell-cell communication analysis}
\description{
Visualize the results of cell-cell communication analysis based on CellPhoneDB.
}
\author{
\strong{Maintainer}: Siyuan Huang \email{huangsiyuan1001@163.com}

}
\keyword{internal}
